moving components to relevant paths according to their usage
