+++ client/components/index.js
@@ -25,8 +25,8 @@ export { default as ButtonGroup } from './button-group';
 export { default as ButtonIcon } from './button-icon';
 export { default as Copy } from './copy';
 export { default as DataViewer } from './data-viewer';
-export { default as DetailList } from './detail-list';
 export { default as DateRangePicker } from './date-range-picker';
+export { default as DetailList } from './detail-list';
 export { default as DomainNavigation } from './domain-navigation';
 export { default as ErrorMessage } from './error-message';
 export { default as FeatureFlag } from './feature-flag';
@@ -40,8 +40,7 @@ export { default as NavigationLink } from './navigation-link';
 export { default as NewsModal } from './news-modal';
 export { default as NoResults } from './no-results';
 export { default as NotificationBar } from './notification-bar';
+export { default as SettingsFooter } from './settings-footer';
+export { default as SettingsToggle } from './settings-toggle';
 export { default as TextInput } from './text-input';
-export { default as SettingsFooter } from './settings-modal/components/settings-footer';
-export { default as SettingsToggle } from './settings-modal/components/settings-toggle';
-export { default as SettingsModal } from './settings-modal';
 export { default as WorkflowGrid } from './workflow-grid';
