+++ client/containers/settings-modal/component.vue
@@ -20,9 +20,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import FlexGrid from '../flex-grid';
-import FlexGridItem from '../flex-grid-item';
-import { SettingsWorkflowHistory } from '../../containers';
+import { FlexGrid, FlexGridItem } from '~components';
+import { container as SettingsWorkflowHistory } from '../settings-workflow-history';
 import SettingsDateFormat from './components/settings-date-format';
 import SettingsHeader from './components/settings-header';
 import SettingsList from './components/settings-list';
