+++ client/containers/settings-modal/index.js
@@ -1,0 +1,7 @@
+import Component from './component';
+
+// TODO - refactor App.vue to reference props from vuex store
+// const container = Connector('SettingsModal', Component);
+const container = Component;
+
+export { container };
